x/ewasm/types/params.go
package types

import (
	"fmt"

	paramtypes "github.com/cosmos/cosmos-sdk/x/params/types"
)

// Parameter store key
var (
	DefaultEnableEwasm       = false
	ParamStoreKeyEnableEwasm = []byte("EnableEwasm")
)

// ParamKeyTable returns the parameter key table.
func ParamKeyTable() paramtypes.KeyTable {
	return paramtypes.NewKeyTable().RegisterParamSet(&Params{})
}

// NewParams creates a new Params object
func NewParams(
	enableEwasm bool,
) Params {
	return Params{
		EnableEwasm: enableEwasm,
	}
}

func DefaultParams() Params {
	return Params{
		EnableEwasm: DefaultEnableEwasm,
	}
}

// ParamSetPairs returns the parameter set pairs.
func (p *Params) ParamSetPairs() paramtypes.ParamSetPairs {
	return paramtypes.ParamSetPairs{
		paramtypes.NewParamSetPair(ParamStoreKeyEnableEwasm, &p.EnableEwasm, validateBool),
	}
}

func validateBool(i interface{}) error {
	_, ok := i.(bool)
	if !ok {
		return fmt.Errorf("invalid parameter type: %T", i)
	}

	return nil
}

func (p Params) Validate() error {
	return validateBool(p.EnableEwasm)
}
